Refactor Jupyter Connection
