+++ src/kernels/jupyter/connection/jupyterConnection.unit.test.ts
@@ -29,7 +29,7 @@ import { DataScience } from '../../../platform/common/utils/localize';
 import { EXTENSION_ROOT_DIR } from '../../../platform/constants.node';
 import { ServiceContainer } from '../../../platform/ioc/container';
 import { IServiceContainer } from '../../../platform/ioc/types';
-import { JupyterConnectionWaiter } from '../launcher/jupyterConnection.node';
+import { JupyterConnectionWaiter } from '../launcher/jupyterConnectionWaiter.node';
 import { noop } from '../../../test/core';
 use(chaiAsPromised);
 suite('Jupyter Connection', async () => {
@@ -167,7 +167,7 @@ suite('JupyterConnection', () => {
         when(serviceContainer.get<IConfigurationService>(IConfigurationService)).thenReturn(instance(configService));
     });
 
-    function createConnectionWaiter(cancelToken?: CancellationToken) {
+    function createConnectionWaiter() {
         return new JupyterConnectionWaiter(
             launchResult,
             notebookDir,
@@ -175,8 +175,7 @@ suite('JupyterConnection', () => {
             // eslint-disable-next-line @typescript-eslint/no-explicit-any
             getServerInfoStub as any,
             instance(serviceContainer),
-            undefined,
-            cancelToken
+            undefined
         );
     }
     test('Successfully gets connection info', async () => {
@@ -184,7 +183,7 @@ suite('JupyterConnection', () => {
         const waiter = createConnectionWaiter();
         observableOutput.next({ source: 'stderr', out: 'Jupyter listening on http://123.123.123:8888' });
 
-        const connection = await waiter.waitForConnection();
+        const connection = await waiter.ready;
 
         assert.equal(connection.localLaunch, true);
         assert.equal(connection.baseUrl, expectedServerInfo.url);
@@ -195,7 +194,7 @@ suite('JupyterConnection', () => {
         (<any>dsSettings).jupyterLaunchTimeout = 10;
         const waiter = createConnectionWaiter();
 
-        const promise = waiter.waitForConnection();
+        const promise = waiter.ready;
 
         await assert.isRejected(promise, DataScience.jupyterLaunchTimedOut);
     });
@@ -203,7 +202,7 @@ suite('JupyterConnection', () => {
         const exitCode = 999;
         const waiter = createConnectionWaiter();
 
-        const promise = waiter.waitForConnection();
+        const promise = waiter.ready;
         childProc.emit('exit', exitCode);
         observableOutput.complete();
